fix:we'd better translate class to another in util class
